Only set defaults when they're not provided to the component
It may be that some properties are left unset.  We provide sensible
defaults in those cases.  It is currently desired to set these
properties in the init for complex objects, instead of defining them
in the class template.  Being in line with these hints doesn't make
the component easier to define, but staying with current best
practices makes things easier to consume afterwards.