Also enable searchDebounceTime on bodyless DataTable invocation
This is a missing feature from #15.

When invoking the databatable as:

    <DataTable ... @searchDebounceTime={{1000}} />

No timeout would occur.  The default value of 2000ms would also not be
used.

This fixes <DataTable ... /> invocation (without a body) to also pass
through the debounce time.  The component will thus predictably abide
the searchDebounceTime parameter and correctly use the default of
2000ms.