Named blocks for DataTable component
This component is the first to receive support for named blocks.  This
allows a user to override only part of the component.

The pattern that arrizes in the component is feasible to stare through
and so we hope this may be a construction that we could iterate on for
the other components too.