Add support for view argument to limit arguments to pass
This is a wrapper object containing:
- size
- page
- filter
- sort
- isLoading
- updatePage
- updatePageSize
- updateFilter
- updateSort

It is an opaque object from the user's point of view.  All current
arguments take precedence making it easy to override.  The default case
becomes shorter which is great to show what's special about the specific
Data Table at hand and thus lowers mental overhead for the default case.